pyproject.toml
[project]
name = "task-categorization-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
]
